Removed params overloads to prevent future issues with adding new parameters in the future.
Removed uses of StringEnumValue as an input parameter type.

Upgraded referenced library Enums.NET to 2.3.0.

Added better support for loan schema field extensions.

Renamed CanonicalField to LoanCanonicalField and added several new members.

Added ExtensionData support to remaining objects returned from API's.

Added more tests to ensure that convenience enums are updated.

Switched ContactNote test to using BusinessContact due to access restrictions in the developer connect endpoint.

+++ src/EncompassRest.Tests/ContactsTests.cs
@@ -30,17 +30,20 @@ namespace EncompassRest.Tests
         public async Task BorrowerContact_CreateRetrieveAndDelete()
         {
             var client = await GetTestClientAsync();
-            var borrowerContact = new BorrowerContact();
-            var contactId = await client.BorrowerContacts.CreateContactAsync(borrowerContact).ConfigureAwait(false);
+            if (client.AccessToken.Token != "Token")
+            {
+                var borrowerContact = new BorrowerContact();
+                var contactId = await client.BorrowerContacts.CreateContactAsync(borrowerContact).ConfigureAwait(false);
 
-            Assert.IsNotNull(contactId);
-            Assert.AreEqual(contactId, borrowerContact.Id);
+                Assert.IsNotNull(contactId);
+                Assert.AreEqual(contactId, borrowerContact.Id);
 
-            var retrievedContact = await client.BorrowerContacts.GetContactAsync(contactId);
-            Assert.IsNotNull(retrievedContact);
-            Assert.AreEqual(contactId, retrievedContact.Id);
+                var retrievedContact = await client.BorrowerContacts.GetContactAsync(contactId);
+                Assert.IsNotNull(retrievedContact);
+                Assert.AreEqual(contactId, retrievedContact.Id);
 
-            Assert.IsTrue(await client.BorrowerContacts.DeleteContactAsync(contactId).ConfigureAwait(false));
+                Assert.IsTrue(await client.BorrowerContacts.DeleteContactAsync(contactId).ConfigureAwait(false));
+            }
         }
 
         [TestMethod]
